requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python-dotenv==1.0.0
psycopg==3.1.8

# for test
# pytest==7.2.0
# django-pytest==0.2.0

# for fill_db
# Faker==19.2.0
